requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
websockets==14.2
openai==1.63.2
python-dotenv==1.0.1
aiohttp==3.11.12
transformers>=4.48.0
torch>=2.2.0
accelerate>=0.26.0
